API/src/apis/models_requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
Werkzeug==3.0.1
pymongo==4.0.1
python-dotenv==0.19.1
flask-hashing==1.1
pyjwt==2.3.0
tensorflow==2.16.1
scikit-learn==1.5.0
pandas==2.2.2